requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==41.0.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
idna==3.4
importlib-metadata==6.7.0
influxdb==5.3.1
influxdb-client==1.37.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
mysql-connector-python==8.2.0
numpy==1.26.2
paho-mqtt==1.6.1
pandas==2.1.3
Pillow==10.0.0
protobuf==4.21.12
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3
reactivex==4.0.4
requests==2.31.0
setuptools==68.2.2
six==1.16.0
svgwrite==1.4.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.6
Werkzeug==2.3.6
zipp==3.15.0
